requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
checkout==0.2.0
Django==3.0.8
django-cors-headers==3.6.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-filters==0.3.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
iniconfig==1.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
Pillow==7.0.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytz==2020.1
requests==2.23.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.1.9
wrapt==1.12.1
zipp==2.2.0
